Everything is working.
I found that I was trying to overcomplicate how the box size was
calculated. I realized that if no points matched the box it would be
done very quickly. so for now I picked a hard coded box size and
increment when no match is found. I tested the worst case scenario given
our data and found that it didn't exceed 5 iterations of the expansion.

I don't believe this will be the fastest solution by a long shot.
However I feel that it is very close to 100% accurate and still speedy
enough to be viable.